EGC Challenge from Grenoble (2017)
This is the final report of my group work from University of Tours. We worked on the EGC Challenge from Grenoble (2017) to develop our data skills. In this perspective, we developed a Data Warehouse to store the data and analyze it. We also conducting a data analysis, lots of feature engineering jobs and tested uni-label and multi-label classifications approaches with different sorts of ML tools.